# Updated HPXML validation