--- a/Gemfile
+++ b/Gemfile
@@ -2,9 +2,7 @@
 
 source 'http://rubygems.org'
 
-gem 'nokogiri', '~> 1.10'
 gem 'oga'
-gem 'schematron-nokogiri'
 gem 'rake'
 gem 'minitest', '5.15'
 gem 'minitest-reporters'