## Pull Request Description

Closes https://github.com/NREL/OpenStudio-HPXML/issues/1100. Switches to performing XSD/Schematron validation via the OS SDK, which is being added for OS v3.5. We now do _full_ XSD schema validation (previously we were only performing limited checks) and can remove a lot of workaround code.

This also provides some nice runtime improvements (results for base.xml):

  | Previous [s] | Now [s] | Change [s]
-- | -- | -- | --
XSD validation | 0.23 | 0.02 | -0.21
Schematron validation | 0.53 | 0.28 | -0.25
Total validation | 0.76 | 0.30 | -0.46

~Ideally the schematron validation would be even faster if we can have OS avoid creating the xslt for every HPXML validation.~ Edit: Turns out supplying the xslt directly is only ~0.05 seconds faster, so it's not a bottleneck.

## Checklist

PR Author: Check these when they're done. Not all may apply. ~~strikethrough~~ and check any that do not apply. 

PR Reviewer: Verify each has been completed.

- [ ] Schematron validator (`EPvalidator.xml`) has been updated
- [ ] Sample files have been added/updated (via `tasks.rb`)
- [ ] Unit tests have been added/updated (e.g., `HPXMLtoOpenStudio/tests`)
- [ ] Documentation has been updated
- [ ] Changelog has been updated
- [ ] `openstudio tasks.rb update_measures` has been run
- [ ] Checked the code coverage report on CI
- [ ] No unexpected changes to simulation results of sample files